travel_planner/requirements.txt
Flask  
requests  
python-dotenv  
google-auth  
google-auth-oauthlib  
google-auth-httplib2  
google-api-python-client  